requirements.txt
# Core AI/ML dependencies
openai==1.54.0
chromadb==0.4.24

# Configuration
python-dotenv==1.0.1

# Document processing
python-docx==1.1.0
PyMuPDF==1.23.26

# Web framework
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
sse-starlette==2.0.0

# UI
streamlit==1.31.1

# HTTP client
requests==2.31.0

# Process management
honcho==1.1.0

# Security
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# File type detection
python-magic==0.4.27
